Moved pool tests to new project.
