+++ AElf.OS.Core.Tests/OSCoreNetworkServiceTestBase.cs
@@ -2,11 +2,6 @@
 
 namespace AElf.OS
 {
-    public class OSCoreTestBase: AElfIntegratedTest<OSCoreTestAElfModule>
-    {
-    
-    }
-    
     public class OSCoreNetworkServiceTestBase: AElfIntegratedTest<NetworkServiceTestModule>
     {
