+++ AElf.OS.Network.Grpc.Tests/GrpcNetworkTestBase.cs
@@ -1,0 +1,9 @@
+using AElf.TestBase;
+
+namespace AElf.OS.Network
+{
+    public class GrpcNetworkTestBase : AElfIntegratedTest<GrpcNetworkTestModule>
+    {
+        
+    }
+}
