+++ AElf.OS.Network.Grpc.Tests/GrpcNetworkTestModule.cs
@@ -1,0 +1,12 @@
+using AElf.Modularity;
+using AElf.OS.Network.Grpc;
+using Volo.Abp.Modularity;
+
+namespace AElf.OS.Network
+{
+    [DependsOn(typeof(OSCoreTestAElfModule), typeof(GrpcNetworkModule))]
+    public class GrpcNetworkTestModule : AElfModule
+    {
+        
+    }
+}
